requirements.txt
azure-functions
tweepy
profanity-check
scikit-learn==0.20.3
quantumrand
better-profanity